requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# Database
supabase==2.9.0

# Environment Variables
python-dotenv==1.0.0

# API Requests
requests==2.31.0

# Stock Market Data
yfinance>=0.2.48

# Data Processing (install later when needed)
# pandas==2.1.4
# numpy==1.26.2
